rearrange the nuklear opengl rendering code a bit
